+++ crates/precompiles/src/contracts/stablecoin_exchange/orderbook.rs
@@ -273,7 +273,7 @@ impl Orderbook {
                 orderbook_slot + offsets::ORDERBOOK_BEST_BID_TICK_OFFSET,
             )
             .expect("TODO: handle error")
-            .to::<i16>();
+            .to::<u16>() as i16;
 
         let best_ask_tick = storage
             .sload(
@@ -281,7 +281,7 @@ impl Orderbook {
                 orderbook_slot + offsets::ORDERBOOK_BEST_ASK_TICK_OFFSET,
             )
             .expect("TODO: handle error")
-            .to::<i16>();
+            .to::<u16>() as i16;
 
         Self {
             base,
